requirements.txt
asgiref==3.8.1
boto3==1.35.25
botocore==1.35.25
cffi==1.17.1
cryptography==43.0.1
django==4.2.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3